requirements.txt
altgraph==0.17.4
aniso8601==9.0.1
attrs==23.2.0
autopep8==2.1.0
blinker==1.8.2
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
dropbox==11.36.2
fdb==2.0.2
Flask==3.0.3
flask-restx==1.3.0
Flask-SQLAlchemy==3.1.1
future==1.0.0
greenlet==3.0.3
idna==3.7
importlib_resources==6.4.0
itsdangerous==2.2.0
Jinja2==3.1.4
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
MarkupSafe==2.1.5
mysql-connector-python==8.4.0
packaging==24.0
pefile==2023.2.7
pillow==10.3.0
ply==3.11
plyer==2.1.0
protobuf==5.26.1
psutil==5.9.8
pycodestyle==2.11.1
pyinstaller==6.6.0
pyinstaller-hooks-contrib==2024.6
pypiwin32==223
PyQt5-Qt5==5.15.2
PyQt5-sip==12.13.0
pytz==2024.1
pywin32==306
pywin32-ctypes==0.2.2
referencing==0.35.1
requests==2.31.0
rpds-py==0.18.1
setuptools==69.5.1
six==1.16.0
SQLAlchemy==2.0.30
stone==3.3.3
typing_extensions==4.11.0
urllib3==2.2.1
Werkzeug==3.0.3
